% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3Dplotting.R
\name{decorate_png}
\alias{decorate_png}
\title{Decorates a png with title, caption and a legend from an image using magick package}
\usage{
decorate_png(
  file = NULL,
  title = "",
  caption = "",
  size = 60,
  legend_file = NULL,
  overwrite = F
)
}
\arguments{
\item{file}{File to decorate}

\item{title}{Title string}

\item{caption}{Caption string}

\item{size}{Size of Title in pix. Caption is 0.7xsize}

\item{legend_file}{Image file with legend will be put on the right}

\item{overwrite}{Should orig file be overwritten? Default: FALSE}
}
\value{
called for the side effect
}
\description{
Decorates a png with title, caption and a legend from an image using magick package
}
