% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3Dplotting.R
\name{mp4_to_gif}
\alias{mp4_to_gif}
\title{Converts mp4 to gif}
\usage{
mp4_to_gif(mp4file, vfilter = "scale=240:-1:flags=lanczos")
}
\arguments{
\item{vfilter}{}
}
\value{
called for the side effect
}
\description{
Converts mp4 to gif
}
