% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3Dplotting.R
\name{renderVideos}
\alias{renderVideos}
\title{Renders mp4 and gif videos from array of image files}
\usage{
renderVideos(
  files,
  rewindLoop = F,
  loop = 0,
  framerate = 16,
  makegif = FALSE,
  gifShrinkage = 4,
  ...
)
}
\arguments{
\item{files}{Array of files used to create animation}

\item{rewindLoop}{Should animation go forth and back? Default: FALSE}

\item{loop}{How often should video be looped? Default: 0}

\item{framerate}{Framerate. Default: 16}

\item{makegif}{Should also a gif be created with gifski? Default: FALSE}

\item{gifShrinkage}{Factor by which gif should be shrinked}
}
\value{
mp4 filename
}
\description{
Renders mp4 and gif videos from array of image files
}
\examples{
brick2movie(br, eleRast = globalDEM0.5deg, gaussianSmoothSigma = 1, renderVideo = F, renderSphere=T, renderOcean=T, pal=rev(pal_MR$divViriMagma), renderCountries = T, thetaStartEnd = c(0,360), nSubSteps = 2, nRounds = 10) \%>\% renderVideos()
}
