% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3Dplotting.R
\name{rs_surface}
\alias{rs_surface}
\title{Title}
\usage{
rs_surface(
  elmat = volcano,
  coloring = volcano,
  img_overlay1 = NULL,
  img_overlay2 = NULL,
  water = FALSE,
  wateralpha = 0.5,
  alpha_over = 0.75,
  texture = "bw",
  zscale = 250,
  shadowintens = 0.8,
  shadowReachFactor = 1,
  windowsize = c(1200, 800),
  zoom = 0.5,
  valRange = range(coloring, na.rm = TRUE, finite = TRUE),
  plot = TRUE,
  pal = palette(),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
