% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3Dplotting.R
\name{testAllRenderOptions}
\alias{testAllRenderOptions}
\title{Test function to show all rendering options}
\usage{
testAllRenderOptions(br = br)
}
\value{
Called for the side effect
}
\description{
Sphere or not sphere, ocean or not etc.
}
\examples{
## Make sure you have the right palette() in place
 testAllRenderOptions()
}
