% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3Dplotting.R
\name{wavesAnim}
\alias{wavesAnim}
\title{Little demo function to create an animation of a wave}
\usage{
wavesAnim()
}
\value{
called for the side effect of the mp4 created
}
\description{
Little demo function to create an animation of a wave
}
\examples{
wavesAnim()
}
